requirements.txt
asgiref==3.7.2
Django==5.0.1
djangorestframework==3.14.0
gensim==3.6.0
numpy==1.26.3
PyPDF2==3.0.1
pytz==2023.3.post1
scipy==1.11.4
six==1.16.0
smart-open==6.4.0
sqlparse==0.4.4
typing_extensions==4.9.0

